pyproject.toml
[project]
name = "fast-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[all] (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "fastapi-cli (>=0.0.10,<0.0.11)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "alembic (>=1.16.5,<2.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "setuptools (>=80.9.0,<81.0.0)",
    "aioredis (==2.0.1)",
    "redis (>=6.4.0,<7.0.0)",
    "fastapi-mail (>=1.5.0,<2.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "asgiref (>=3.9.1,<4.0.0)",
    "st (>=0.0.8,<0.0.9)",
]

[tool.poetry]
packages = [{include = "fast_api", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
